% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WID_BC.R
\name{WID_BC}
\alias{WID_BC}
\title{WID_BC}
\usage{
WID_BC(beta)
}
\arguments{
\item{beta}{Methylation beta matrix object (Illumina Methylation EPIC array)}
}
\value{
Returns a vector object with the WID-BC-index
}
\description{
THE WID_BC function uses a methylation beta matrix and returns the WID-BC-index.
Authors: Barrett J., Herzog C. et al., 2021
}
